Copy all children of the first parsed XML file
Copy all of the children of the first parsed XML file, this avoids
missing parts of the first parsed XML file.

Most XML files have all NODE info as children of the first NODE,
therefore when the first NODE was the first child, no issues occurred.
However, many XML files first child is a SCRIPT node. For these the
NODE schema tree was not copied.