merge_nodes: Use xmlAddChild instead of setting children
With newer releases of libxml2, merge_nodes()'s direct setting of the
xmlNode->children resulted in the tree not being formed in the way libxml2
expected, resulting in the xmlCopyDoc() of sch_dump_xml() segfaulting while
performing a copy of nodes.

To resolve this, the parent node is now passed to merge_nodes(), allowing
the use xmlAddChild() instead of a combination of xmlAddSibiling() and
direct setting of xmlNode->children.