sch_gnode_to_json: Handle NULL values
An apteryx_provide callback may return NULL, resulting in the tree returned
by apteryx_get_tree() containing a NULL value. sch_gnode_to_json currently
does not handle this and attempts to strdup(NULL), resulting in a segfault.

Resolve this by handling this case, and replacing the NULL with an empty
string.