Automatically import/include required modules
But do not include their nodes in the output unless they
are required for the module being processed.
Make sure namespaces are correct for all nodes.
Store namespace and prefix of the module with the model
organisation etc as we cannot tell what the prefix is
later if the nodes are all in the default namespace.