Enhance the handling of Netconf xpaths
This change adds support for the syntax //field or /*/field in an xpath.
The syntax allows the user to specify one or more levels of the tree path.
For example given a netconf get request with the xpath of "/test/animals/*/colour"
or "/test/animals//colour" the resulting reply will return the XML tree:-
  <test xmlns="http://test.com/ns/yang/testing">
    <animals>
      <animal>
        <name>dog</name>
        <colour>brown</colour>
      </animal>
      <animal>
        <name>mouse</name>
        <colour>grey</colour>
      </animal>
      <animal>
        <name>parrot</name>
        <colour>blue</colour>
      </animal>
    </animals>
  </test>