Check for bool type differently
A scheme node that is bool may have more than two nodes which can
be WATCH or PROVIDE which use those bool values.
If this is the case then those values will be returned as "true"
or "false" instead of json bool because xmlChildElementCount does
not ignore the irrelevant nodes.

Fix this issue by implementing a new function to determine
if a scheme is a bool.