Fix two minor problems getting report-all defaults
When the flag SCH_F_ADD_DEFAULTS is set and there are no children
of the child and there is also no schema children, then delete the
child node from the result tree.

When traversing a tree with the flag SCH_F_FILTER_RDEPTH set,
then the starting depth should be 1 reflecting that the schema is
passing its children.