Add support for RFC6020 for "if-feature", "when" and "must" statements
The "if-feature statement is defined in RFC6020 section 7.18.2
The "when" statement is defined in RFC6020 section 7.19.5
The "must" statement is defined in RFC6020 section 7.5.3

These statements are now added as additional attributes to the
generated XML files by pyang-apteryx-xml.py.

The statements are mostly of XPATH 1.0 syntax so I have incorporated
a xpath parser into the code.