Add support for loading and unloading models and model features
This changes allows loaded models to be unloaded if not required and
reloaded at a later point if  required again. Features can now also be
added or removed dynamically from a loaded model.

Support from the yang-library data model has been shifted from
apteryx-rest to this repository, so that apteryx-netconf can use the
common code.