Fixes for with-defaults processing
In _merge_gnode_nodes, don't continue when the child1 node has no
children. This is to allow the recursive call which, while it will have
no effect in the loop over node1's children, may have an effect in the
loop over node2's children.

Change schemas to allow the new tests in apteryx-netconf. The test5 schema
is meant to mimic the structure of the interfaces/ethernet/poe models
in openconfig.